+++ ZoomSDKSample/ZoomSDKSample/UI Interface/ZMSDKSettingWindowController.m
@@ -698,4 +698,10 @@
             break;
     }
 }
+- (void)showWindow:(id)sender
+{
+    [self updateAudioTab];
+    [self updateVideoTab];
+    [self relayoutWindowPosition];
+}
 @end